% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountEvents.R
\name{count_events}
\alias{count_events}
\title{Count events in a data frame using two variable}
\usage{
count_events(data, var1, var2)
}
\arguments{
\item{data}{A data frame or Tibble containing the dataset to analyze, thus data.}

\item{var1}{The first variable used for grouping and event counting, thus var1.}

\item{var2}{The second variable used for grouping and event counting, thus var2.}
}
\value{
A data frame containing counts of events for each unique combination of
the specified variables.
}
\description{
This function takes a data frame or Tibble, and two variable names as arguments.
It performs data validation, filtering out rows with missing values.
counts the number of events in the data, grouped by the two specified variables, and returns the result as a data frame.
}
\examples{
# Example usage:
library(dplyr)
my_data <- data.frame(var1 = c(NA, "A", "B", "B", "A", "B"), var2 = c("X", "Y", "X", "X", NA, "Y"))
my_result <- count_events(my_data, var1, var2)

}
